test(#74): app.jsとserver.jsの分離によるテスト環境修正
- Expressアプリとサーバーインスタンスを分離
- Supertestでapp.address is not a function エラーを解決
- テスト用にapp.jsからExpressアプリオブジェクトをエクスポート
